Move away from git lfs
Remove wav files and replace with compressed npz

Apparently github has a monthly bandwidth limit for Git LFS.
That's reasonable and fair.
What I find unreasonable is that the bandwidth usage by
GH Actions workflows is also counted against the monthly
quota:
https://docs.github.com/en/repositories/working-with-files/managing-large-files/about-storage-and-bandwidth-usage#tracking-storage-and-bandwidth-use

After three workflow runs (4 jobs each), this limit was
reached.

Apparently I'm not the first to fall for this:
https://github.com/actions/checkout/issues/165
https://github.com/actions/checkout/issues/834
https://github.com/orgs/community/discussions/26775

To name a few