Roll scipy back to v1.15.2
Previous commit 00703f32364 mistakenly updated it to v1.16.3
which seems to be where the regression is coming from:

https://github.com/nocarryr/lupy/runs/59043827174

compared to

https://github.com/nocarryr/lupy/runs/59043933175

Remember that in this case, we're looking at things backwards.
Trying to find the regression in main by looking at it from
a PR, so no performance improvement means we've found the
regression (as opposed to increased perf).